requirements.txt
youtube_transcript_api
streamlit~=1.40.1
google-generativeai
python-dotenv~=1.0.1
pathlib~=1.0.1
pygments~=2.18.0
pip~=24.2
attrs~=24.3.0
wheel~=0.44.0
openssl~=3.0.15
pillow~=10.4.0
setuptools~=75.1.0
packaging~=24.2
zipp~=3.20.2